Move golem.core.hardware functions to golem.hardware Move golem.core.hardware.HardwarePresets to golem.hardware.presets Move golem.config.presets.HardwarePresetsMixin to golem.hardware.presets
